+++ src/main/java/com/nincraft/ninbot/components/stream/StreamingMemberRepository.java
@@ -1,4 +1,4 @@
-package com.nincraft.ninbot.components.twitch;
+package com.nincraft.ninbot.components.stream;
 
 import org.springframework.data.repository.CrudRepository;
