+++ src/main/resources/lang_pt.properties
@@ -46,7 +46,7 @@ command.subscribe.description.text=Subscribes you to a game for game gathering e
 command.unsubscribe.description.text=Unsubscribes you from a game for game gathering events
 command.subscribe.norolefound=Could not find the role "%s", contact an admin
 command.trivia.description.text=Starts/Stops trivia
-command.twitch.description.text=Access Twitch related commands
+command.stream.description.text=Access Stream related commands
 command.countdown.help.text=Use "@Ninbot countdown YYYY-MM-DD CountdownName" to setup a countdown. It will be announced every day leading up to the event.
 command.countdown.list.starttime="Start Time\: "
 command.countdown.list.footer="All times are shown in "
@@ -64,7 +64,7 @@ command.pathogen.description.text=Displays the user's infection level. Use "@Nin
 #Listeners
 listener.topic.updated.withpermission=%s updated topic to %s
 listener.topic.update.nopermission=Topic updated to %s
-listener.twitch.announce=%s is streaming %s! Check them out at %s
+listener.stream.announce=%s is streaming %s! Check them out at %s
 listener.dad.imcontraction=I'm
 listener.dad.imnocontraction=im
 listener.dad.hi=Hi
@@ -101,7 +101,7 @@ command.info.title=Ninbot Info
 command.info.git.hash=Git Hash
 command.info.uptime=Uptime
 command.event.help.text=Use "@Ninbot events plan" to add an event to the schedule\nParameters: @Ninbot events plan "Event Name" StartTime GameName\nNote: event name must be in quotes if it is longer than one word\nEvent times are in GMT -6, formatted "2019-01-31T12:00:00-06:00" for January 31st 2019 at noon
-command.twitch.help.text=Access Twitch related commands\n"twitch announce" toggles your Twitch going live announcement
+command.stream.help.text=Access Stream related commands\n"stream announce" toggles your Stream going live announcement
 command.trivia.help.text=Sub commands\nstart [ID] - starts trivia in a channel with an optional category ID\nstop - stop trivia in a channel\ncategories - list available trivia categories\nscore - shows your current score\nleaderboard - shows the top 5 players
 command.lose.help.text=lose @Username
 command.win.help.text=win @Username
